added javadoc, updated spec

+++ airbyte-integrations/connectors/destination-snowflake/src/main/java/io/airbyte/integrations/destination/snowflake/SnowflakeInternalStagingConsumerFactory.java
@@ -31,11 +31,20 @@ import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * Snowflake Internal Staging consists of 4 main parts
+ *
+ * CREATE STAGE @TEMP_STAGE_NAME -- Creates a new named internal stage to use for loading data from
+ * files into Snowflake tables and unloading data from tables into files PUT
+ * file://local/<file-patterns> @TEMP_STAGE_NAME. --JDBC Driver will upload the files into stage
+ * COPY FROM @TEMP_STAGE_NAME -- Loads data from staged files to an existing table.
+ * DROP @TEMP_STAGE_NAME -- Drop temporary stage after sync
+ */
 public class SnowflakeInternalStagingConsumerFactory {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(SnowflakeInternalStagingConsumerFactory.class);
 
-  private static final int MAX_BATCH_SIZE_BYTES = 1024 * 1024 * 1024 / 4; // 256mb
+  private static final long MAX_BATCH_SIZE_BYTES = 1024 * 1024 * 1024 / 4; // 256mb
 
   public static AirbyteMessageConsumer create(final Consumer<AirbyteMessage> outputRecordCollector,
                                               final JdbcDatabase database,
